API-Flask/requirements.txt
Flask==2.1.2
Flasgger==0.9.5
PyYAML==6.0
requests==2.27.1
vercel-flask==0.0.0